% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumer-price-index.R
\docType{data}
\name{cpi}
\alias{cpi}
\title{Consumer Price Index for South Africa.}
\format{A data frame with 3 variables:
\describe{
 \item{year}{}
 \item{month}{}
 \item{cpi}{Official inflation rate.}
}}
\usage{
cpi
}
\description{
Consumer Price Index for South Africa.
}
\keyword{datasets}
