% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpi-weights.R
\docType{data}
\name{cpi_weights}
\alias{cpi_weights}
\title{Consumer Price Weights for South Africa CPI Basket}
\format{
A data frame with 6 variables:
\describe{
 \item{coicop_code}{A code used to group the indicator products into categories.}
 \item{product_code}{An identifier for the indicator products.}
 \item{indicator_product}{The indicator product to which a weight is assigned.}
 \item{provincial_baskets}{Description of which provincial baskets include the indicator product.}
 \item{total_country_weight}{}
 \item{headline_weight}{}
}
}
\usage{
cpi_weights
}
\description{
Consumer Price Weights for South Africa CPI Basket
}
\keyword{datasets}
