% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialingcodes.R
\docType{data}
\name{dialing_codes}
\alias{dialing_codes}
\title{Dialing codes for cities in South Africa}
\format{
A data frame with 2 columns:
\describe{
 \item{city}{}
 \item{codes}{}
}
}
\source{
\url{https://www.rebtel.com/en/international-calling-guide/phone-codes/south-africa/}
}
\usage{
dialing_codes
}
\description{
Dialing codes for cities in South Africa
}
\keyword{datasets}
