% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuel-price.R
\docType{data}
\name{fuel_price}
\alias{fuel_price}
\title{Fuel Price history for South Africa}
\format{
A data frame with 4 columns:
\describe{
 \item{year}{}
 \item{month}{}
 \item{day}{}
 \item{date}{}
 \item{region}{}
 \item{fuel}{}
 \item{price}{Prices are in cents per litre for all fuels except Liquefied Petroleum Gas (LPG), where the price is in cents per kilogram.}
}
}
\usage{
fuel_price
}
\description{
Fuel Price history for South Africa
}
\keyword{datasets}
