% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuel-price-basic.R
\docType{data}
\name{fuel_price_basic}
\alias{fuel_price_basic}
\title{Basic Fuel Price history for South Africa}
\format{
A data frame with 4 columns:
\describe{
 \item{year}{}
 \item{month}{}
 \item{fuel}{}
 \item{price}{Prices are in cents per litre.}
}
}
\source{
\url{http://www.energy.gov.za/files/esources/petroleum/petroleum_arch.html}
}
\usage{
fuel_price_basic
}
\description{
Basic Fuel Price history for South Africa
}
\keyword{datasets}
