% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuel-price-brent-crude.R
\docType{data}
\name{fuel_price_brent_crude}
\alias{fuel_price_brent_crude}
\title{Price history for Brent Crude Oil}
\format{
A data frame with 3 columns:
\describe{
 \item{year}{}
 \item{month}{}
 \item{brent_crude_avg}{Price per barrel in US dollars}
}
}
\source{
\url{http://www.energy.gov.za/files/esources/petroleum/petroleum_arch.html}
}
\usage{
fuel_price_brent_crude
}
\description{
Price history for Brent Crude Oil
}
\keyword{datasets}
