% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuel-price-used-exchange-rate.R
\docType{data}
\name{fuel_price_usd_exchange_rate}
\alias{fuel_price_usd_exchange_rate}
\title{Exchange rate history for the South African Rand against the US Dollar}
\format{
A data frame with 3 columns:
\describe{
 \item{year}{}
 \item{month}{}
 \item{zar_per_usd}{Rands per US dollar}
}
}
\source{
\url{http://www.energy.gov.za/files/esources/petroleum/petroleum_arch.html}
}
\usage{
fuel_price_usd_exchange_rate
}
\description{
Exchange rate history for the South African Rand against the US Dollar
}
\keyword{datasets}
