% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuel-type.R
\docType{data}
\name{fuel_type}
\alias{fuel_type}
\title{Fuel Types}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 13 rows and 2 columns.
}
\usage{
fuel_type
}
\description{
Fuel Types
}
\keyword{datasets}
