% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuelprice.R
\docType{data}
\name{fuelprice}
\alias{fuelprice}
\title{Fuel Price history for South Africa.}
\format{
A data frame with 4 columns:
\describe{
 \item{fuel}{}
 \item{region}{}
 \item{date}{Postal code for Post Office boxes.}
 \item{price}{Prices are in cents per litre for all fuels except Liquefied Petroleum Gas (LPG), where the price is in cents per kilogram.}
}
}
\source{
\url{https://www.sapia.org.za/}
}
\usage{
fuelprice
}
\description{
Fuel Price history for South Africa.
}
\keyword{datasets}
