% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\docType{data}
\name{history}
\alias{history}
\title{History}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 42 rows and 5 columns.
}
\usage{
history
}
\description{
History
}
\keyword{datasets}
