% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\docType{data}
\name{history_tagged}
\alias{history_tagged}
\title{History with Tags}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 40 rows and 7 columns.
}
\usage{
history_tagged
}
\description{
History with Tags
}
\examples{
library(dplyr)

# Extract dates of holidays.
history_tagged \%>\% filter(holiday)
}
\keyword{datasets}
