% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postalcodes.R
\docType{data}
\name{postalcodes}
\alias{postalcodes}
\title{Postal codes for South Africa.}
\format{
A data frame with 4 variables:
\describe{
 \item{area}{}
 \item{suburb}{}
 \item{boxcode}{Postal code for Post Office boxes.}
 \item{streetcode}{Postal code for street addresses.}
}
}
\source{
\url{https://www.postoffice.co.za/contactus/postalcodes.xls}
}
\usage{
postalcodes
}
\description{
Postal codes for South Africa.
}
\keyword{datasets}
