% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postalcodes_geo.R
\docType{data}
\name{postalcodes_geo}
\alias{postalcodes_geo}
\title{Geo location of postal codes for South Africa.}
\format{
A data frame with 3 variables:
\describe{
 \item{code}{Postal code.}
 \item{latitude}{Latitude.}
 \item{lon}{Longitude.}
}
}
\source{
\url{http://www.geonames.org/postal-codes/postal-codes-south-africa.html}
}
\usage{
postalcodes_geo
}
\description{
Geo location of postal codes for South Africa.
}
\keyword{datasets}
