% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/president-speeches.R
\docType{data}
\name{president_speeches}
\alias{president_speeches}
\title{Speeches of the President and Deputy President}
\format{
A data frame with 5 columns:
\describe{
 \item{date}{}
 \item{position}{}
 \item{person}{}
 \item{language}{}
 \item{title}{}
 \item{text}{}
}
}
\source{
\url{http://www.thepresidency.gov.za/speeches/}
}
\usage{
president_speeches
}
\description{
Speeches of the President and Deputy President
}
\keyword{datasets}
