% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saffeR.R
\docType{package}
\name{saffeR}
\alias{saffeR}
\alias{saffeR-package}
\title{saffeR: Things relating to South Africa}
\description{
The saffeR package provides a collection of data and functions for things relevant to South Africa.
}
