% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub-places.R
\docType{data}
\name{sub_places}
\alias{sub_places}
\title{Dataset of Sub-places in South Africa}
\format{
A data frame with 21243 records and 8 variables:
\describe{
  \item{sub_place_code}{}
  \item{sub_place}{}
  \item{municipality_code}{}
  \item{municipality}{}
  \item{district_code}{}
  \item{district}{}
  \item{province_code}{}
  \item{province}{}
}
}
\source{
\url{http://www.statssa.gov.za/census/census_2001/SubPlace.zip}
}
\usage{
data(sub_places)
}
\description{
A dataset containing the the names of South African sub-places (e.g. suburbs).
}
\keyword{datasets}
