% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unemployment-education.R
\docType{data}
\name{unemployment_education}
\alias{unemployment_education}
\title{Unemployment rate of South Africa.}
\format{
A data frame with 4 variables:
\describe{
 \item{education}
 \item{year}{}
 \item{quarter}
 \item{unemployment_rate}{}
}
}
\source{
\url{http://www.statssa.gov.za/publications/P0211/P02113rdQuarter2020.pdf}
}
\usage{
unemployment_education
}
\description{
Unemployment rate of South Africa.
}
\keyword{datasets}
