% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unemployment-gender.R
\docType{data}
\name{unemployment_gender}
\alias{unemployment_gender}
\title{Unemployment rate of South Africa.}
\format{
A data frame with 4 variables:
\describe{
 \item{gender}{}
 \item{year}{}
 \item{quarter}{}
 \item{unemployment rate}{}
}
}
\source{
\url{http://www.statssa.gov.za/publications/P0211/P02113rdQuarter2020.pdf}
}
\usage{
unemployment_gender
}
\description{
Unemployment rate of South Africa.
}
\keyword{datasets}
