% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unemployment-worldbank.R
\docType{data}
\name{unemployment_sa}
\alias{unemployment_sa}
\title{Unemployment rate of South Africa.}
\format{
A data frame with 5 variables:#' \describe{
 \item{country_id}{}
 \item{country_iso3}{country iso3 code}
 \item{country_name}{}
 \item{year}{}
 \item{unemployed_percent}{Unemployment, total (\% of total labor force; modeled ILO estimate)}
}
}
\source{
\url{https://api.worldbank.org/v2/countries/ZA/indicators/SL.UEM.TOTL.ZS?per_page=10000&format=JSON}
}
\usage{
unemployment_sa
}
\description{
Unemployment rate of South Africa.
}
\keyword{datasets}
