% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpi-average-prices.R
\docType{data}
\name{cpi_average_prices}
\alias{cpi_average_prices}
\title{CPI average prices in South Africa}
\format{
A data frame with 4 columns:
\describe{
 \item{product}{}
 \item{year}
 \item{month}
 \item{price}
}
}
\source{

}
\usage{
cpi_average_prices
}
\description{
CPI average prices in South Africa
}
\keyword{datasets}
