% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postalcodes_province.R
\name{postalcodes_province}
\alias{postalcodes_province}
\title{Find province from Postal Code}
\source{
\url{https://en.wikipedia.org/wiki/List_of_postal_codes_in_South_Africa}
}
\usage{
postalcodes_province(code)
}
\description{
Find province from Postal Code
}
