% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saffer.R
\docType{package}
\name{saffer}
\alias{saffer}
\title{saffer: Things relating to South Africa}
\description{
The {saffer} package provides a collection of data and functions for things relevant to South Africa.
}
