% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unemployment-worldbank.R
\docType{data}
\name{unemployment_sa}
\alias{unemployment_sa}
\title{Unemployment rate of South Africa.}
\format{

}
\source{
\url{https://api.worldbank.org/v2/countries/ZA/indicators/SL.UEM.TOTL.ZS?per_page=10000&format=JSON}
}
\usage{
unemployment_sa
}
\description{
Unemployment rate of South Africa.
}
\keyword{datasets}
